+++ apps/backend/api/user/preferences/[module].delete.ts
@@ -1,5 +1,5 @@
 import { eventHandler, getRouterParam, setResponseStatus } from 'h3';
-import { PreferenceService } from '~/modules/preference/preference.service';
+import { PreferenceService } from '~/modules/user/preference.service';
 import { logApiError } from '~/utils/api-logger';
 import { verifyAccessToken } from '~/utils/jwt-utils';
 import {