+++ apps/backend/modules/inspection/inspection-api.service.ts
@@ -2,7 +2,7 @@ import type { UserSession } from '~/utils/jwt-utils';
 
 import { FileStorageService } from '~/modules/file-storage/file-storage.service';
 import { SystemLogService } from '~/modules/system-log/system-log.service';
-import { WelderScoreService } from '~/modules/welder-score/welder-score.service';
+import { WelderScoreService } from '~/modules/welder/welder-score.service';
 import { recordBusinessAuditLog } from '~/utils/audit-log';
 import {
   buildImportRowError,