+++ apps/backend/modules/master-data-rename/index.ts
@@ -1,6 +1,0 @@
-export { MasterDataRenameService } from './master-data-rename.service';
-export type {
-  MasterDataOrphanItem,
-  RenameRequest,
-  RenameResult,
-} from './master-data-rename.service';