+++ apps/backend/modules/route-handlers/auth/login.post.service.ts
@@ -1,6 +1,6 @@
 import { LoginStatusEnum } from '@qgs/shared';
 import { defineEventHandler, getHeader, readBody, setResponseStatus } from 'h3';
-import { AuthService } from '~/modules/auth/auth.service';
+import { AuthService } from '~/modules/user/auth.service';
 import { SystemLogService } from '~/modules/system-log/system-log.service';
 import {
   clearRefreshTokenCookie,